Bug fixes
Fixed how steps 4-6 read and write data when pulling and aggregating mass tracks and spot data .csv files. Also changed step 6 to output track classification results into a csv file instead of a text file.

Finally, changed step 6 to trim the test dataset between min.timepoint and max.timepoint and adjusted the max.timepoint for trimming the training dataset because of small difference in frame interval between imaging runs. This does not change any of the ground-truth manual classifications in the training set, but ensures that the training and test datasets have the same number of data points. In the future, we will have to rewrite step 6 code to replace "number of N/A mass values" as a classification metric to make our classification code robust to differences in imaging times and small differences in frame intervals.