Step 6 update to match time windows in training and test datasets
In training datasets, 6-72 hours post-treatment corresponded to frames 1-381, for a total of 381 frames. In our most recent test datasets, 6-72 hours post-treatment corresponded to frames 24-407, for a total of 384 frames. Here, we're pulling the first 381 frames from the training datasets and frames 24-407 (384 frames) from the test datasets to equalize the duration of the datasets and their relative timing post-treatment, rather than equalizing their number of frames such that the duration of the training and test datasets differed by 30 minutes.